fix(mcp): preserve OAuth reauthorization fallback
