+++ docs/.generated/plugin-sdk-api-baseline/agent-harness.json
@@ -1,1 +1,1 @@
-{"contentHash":"108af57a4b4bc57c70912fc40319876bb6ec994cce59d2b9c991980e18c4f80f","entrypoint":"agent-harness","importSpecifier":"openclaw/plugin-sdk/agent-harness"}
+{"contentHash":"77367f9045d9be79922dd1317dcaa6666799a527afeea0f0ec362e1987dfeefe","entrypoint":"agent-harness","importSpecifier":"openclaw/plugin-sdk/agent-harness"}
