+++ docs/.generated/plugin-sdk-api-baseline/channel-core.json
@@ -1,1 +1,1 @@
-{"contentHash":"dd255204bf11dccfe9d866e7a6725935bfead7378a52ddfa6b60b9a58c0cdf76","entrypoint":"channel-core","importSpecifier":"openclaw/plugin-sdk/channel-core"}
+{"contentHash":"4503457a71010a036864d006a1d424518d70fe3e5a0fedc565e5b8028cec0067","entrypoint":"channel-core","importSpecifier":"openclaw/plugin-sdk/channel-core"}
