+++ docs/.generated/plugin-sdk-api-baseline/channel-message.json
@@ -1,1 +1,1 @@
-{"contentHash":"8bf7d1fd1e21861fd472365e0ab10ca811800e1b0cb9e6aac0e4d6a41613a9a2","entrypoint":"channel-message","importSpecifier":"openclaw/plugin-sdk/channel-message"}
+{"contentHash":"90e637c3f68e33e00fd719a4c6429c619c0ba61b1fc741aff902aa87fb1dcebb","entrypoint":"channel-message","importSpecifier":"openclaw/plugin-sdk/channel-message"}
