+++ docs/.generated/plugin-sdk-api-baseline/core.json
@@ -1,1 +1,1 @@
-{"contentHash":"4c667b7e88cd752c294bcc731bc157c1ccb2a7aa4afc58f4013409191f3a9b28","entrypoint":"core","importSpecifier":"openclaw/plugin-sdk/core"}
+{"contentHash":"698ff6f1452941b8cf6a0f2789885f38cb7c9bcd0ef4a38f7953d27ffdca8b1e","entrypoint":"core","importSpecifier":"openclaw/plugin-sdk/core"}
