+++ docs/.generated/plugin-sdk-api-baseline/discord.json
@@ -1,1 +1,1 @@
-{"contentHash":"027b2280171324b628865fb4cdf9f72c13c059385a2fbec5b650c69d66beecdc","entrypoint":"discord","importSpecifier":"openclaw/plugin-sdk/discord"}
+{"contentHash":"03ba0dcb9f9238961917a183c69448649718af31bce53281bfea06a9017b79ee","entrypoint":"discord","importSpecifier":"openclaw/plugin-sdk/discord"}
