+++ src/agents/mcp-oauth.test.ts
@@ -1,26 +1,19 @@
-import { createHash } from "node:crypto";
 import fs from "node:fs/promises";
-import { createServer, type IncomingMessage, type ServerResponse } from "node:http";
 import path from "node:path";
 import { withTempHome as withBaseTempHome } from "openclaw/plugin-sdk/test-env";
 import { afterEach, beforeEach, describe, expect, it } from "vitest";
 import { vi } from "vitest";
-import type { McpServerConfig } from "../config/types.mcp.js";
-import { handleMcpOAuthCallback } from "../gateway/mcp-oauth-callback.js";
-import { createRequest, createResponse } from "../gateway/server-http.test-harness.js";
 import {
   closeOpenClawStateDatabaseForTest,
   openOpenClawStateDatabase,
 } from "../state/openclaw-state-db.js";
 import { resolveOpenClawStateSqlitePath } from "../state/openclaw-state-db.paths.js";
-import { getFreePort } from "../test-utils/ports.js";
 import {
   operatorMcpOAuthIdentity,
   requesterMcpOAuthIdentity,
   type McpOAuthIdentity,
 } from "./mcp-oauth-identity.js";
 import { createMcpOAuthClientProvider } from "./mcp-oauth-provider.js";
-import { readMcpOAuthPendingAuthorization as readPending } from "./mcp-oauth-store.js";
 import { readMcpOAuthStore, updateMcpOAuthStore } from "./mcp-oauth-store.js";
 import {
   clearMcpOAuthCredentials,
@@ -32,7 +25,6 @@ import {
   resolveMcpOAuthAccessToken,
   startMcpOAuthAuthorization,
 } from "./mcp-oauth.js";
-import { resolveMcpTransportConfig } from "./mcp-transport-config.js";
 
 const authMock = vi.hoisted(() => vi.fn());
 const ROTATED_ACCESS = "gateway-token";
@@ -56,26 +48,6 @@ async function saveAccessToken(identity: McpOAuthIdentity, accessToken: string):
   });
 }
 
-async function runGatewayOAuthCallback(params: {
-  serverName: string;
-  server: McpServerConfig;
-  code: string;
-  state: string;
-}) {
-  const response = createResponse();
-  await handleMcpOAuthCallback(
-    createRequest({
-      path: `/oauth/mcp/callback?code=${params.code}&state=${params.state}`,
-    }),
-    response.res,
-    {
-      config: { mcp: { servers: { [params.serverName]: params.server } } },
-      log: { warn: vi.fn() },
-    },
-  );
-  return response;
-}
-
 function resolvedOAuthConfig(identity: McpOAuthIdentity) {
   return {
     kind: "http" as const,
@@ -96,90 +68,6 @@ async function persistRedirect(provider: ReturnType<typeof createMcpOAuthClientP
   authorizationUrl.searchParams.set("state", "state-1234567890");
   await provider.redirectToAuthorization(authorizationUrl);
   return "REDIRECT" as const;
-}
-
-function sendOAuthJson(response: ServerResponse, body: unknown, status = 200): void {
-  response.writeHead(status, { "content-type": "application/json" });
-  response.end(JSON.stringify(body));
-}
-
-async function readOAuthBody(request: IncomingMessage): Promise<string> {
-  const chunks: Buffer[] = [];
-  for await (const chunk of request) {
-    chunks.push(Buffer.isBuffer(chunk) ? chunk : Buffer.from(chunk));
-  }
-  return Buffer.concat(chunks).toString("utf8");
-}
-
-async function startAuthorizationServer(port: number) {
-  const issuer = `http://127.0.0.1:${port}`;
-  const handleRequest = async (request: IncomingMessage, response: ServerResponse) => {
-    const url = new URL(request.url ?? "/", issuer);
-    if (url.pathname.startsWith("/.well-known/oauth-protected-resource")) {
-      sendOAuthJson(response, { resource: `${issuer}/mcp`, authorization_servers: [issuer] });
-      return;
-    }
-    if (url.pathname === "/.well-known/oauth-authorization-server") {
-      sendOAuthJson(response, {
-        issuer,
-        authorization_endpoint: `${issuer}/authorize`,
-        token_endpoint: `${issuer}/token`,
-        registration_endpoint: `${issuer}/register`,
-        response_types_supported: ["code"],
-        grant_types_supported: ["authorization_code", "refresh_token"],
-        token_endpoint_auth_methods_supported: ["none"],
-        code_challenge_methods_supported: ["S256"],
-      });
-      return;
-    }
-    if (url.pathname === "/register" && request.method === "POST") {
-      const metadata = JSON.parse(await readOAuthBody(request)) as Record<string, unknown>;
-      sendOAuthJson(response, { ...metadata, client_id: "fixture-client" }, 201);
-      return;
-    }
-    if (url.pathname === "/token" && request.method === "POST") {
-      const form = new URLSearchParams(await readOAuthBody(request));
-      const challenge = createHash("sha256")
-        .update(form.get("code_verifier") ?? "")
-        .digest("base64url");
-      if (form.get("code") !== challenge) {
-        sendOAuthJson(response, { error: "invalid_grant" }, 400);
-        return;
-      }
-      sendOAuthJson(response, {
-        access_token: `access-${challenge.slice(0, 8)}`,
-        refresh_token: "fixture-refresh",
-        token_type: "Bearer",
-        expires_in: 3600,
-      });
-      return;
-    }
-    response.writeHead(404).end();
-  };
-  const server = createServer((request, response) => {
-    void handleRequest(request, response).catch((error: unknown) => {
-      response.destroy(error instanceof Error ? error : new Error("OAuth fixture failed"));
-    });
-  });
-  await new Promise<void>((resolve, reject) => {
-    server.once("error", reject);
-    server.listen(port, "127.0.0.1", resolve);
-  });
-  return {
-    issuer,
-    close: () =>
-      new Promise<void>((resolve) => {
-        server.close(() => resolve());
-      }),
-  };
-}
-
-function authorizationCode(authorizationUrl: string): string {
-  const code = new URL(authorizationUrl).searchParams.get("code_challenge");
-  if (!code) {
-    throw new Error("authorization URL omitted the PKCE challenge");
-  }
-  return code;
 }
 
 vi.mock("@modelcontextprotocol/sdk/client/auth.js", () => ({
@@ -962,105 +850,6 @@ describe("MCP OAuth provider", () => {
           OPENCLAW_CONFIG_PATH: undefined,
           OPENCLAW_STATE_DIR: undefined,
         },
-      },
-    );
-  });
-
-  it("resumes authorization after restart, retains failures, and supersedes older starts", async () => {
-    await withTempHome(
-      async () => {
-        const { auth: realAuth } = await vi.importActual<
-          typeof import("@modelcontextprotocol/sdk/client/auth.js")
-        >("@modelcontextprotocol/sdk/client/auth.js");
-        authMock.mockImplementation(realAuth);
-        const fixture = await startAuthorizationServer(await getFreePort());
-        const rawServer = {
-          url: `${fixture.issuer}/mcp`,
-          transport: "streamable-http" as const,
-          auth: "oauth" as const,
-          oauth: {
-            identity: "per-requester" as const,
-            redirectUrl: "https://gateway.example.com/oauth/mcp/callback",
-          },
-        };
-        const config = resolveMcpTransportConfig("fixture", rawServer);
-        if (config?.kind !== "http") {
-          throw new Error("expected HTTP MCP OAuth config");
-        }
-        const identity = requesterIdentity("fixture", config.url, "sender-a");
-        try {
-          const first = await startMcpOAuthAuthorization(identity, config, {});
-          if (first.status !== "redirect") {
-            throw new Error("expected first MCP OAuth redirect");
-          }
-          expect(readMcpOAuthStore(identity.storeKey)).toMatchObject({
-            codeVerifier: expect.any(String),
-            lastAuthorizationUrl: first.authorizationUrl,
-            redirectUrl: first.redirectUrl,
-          });
-          closeOpenClawStateDatabaseForTest();
-          const callbacks = await Promise.all(
-            [0, 1].map(() =>
-              runGatewayOAuthCallback({
-                serverName: "fixture",
-                server: rawServer,
-                code: authorizationCode(first.authorizationUrl),
-                state: first.state,
-              }),
-            ),
-          );
-          expect(callbacks.map(({ res }) => res.statusCode).toSorted((a, b) => a - b)).toEqual([
-            200, 404,
-          ]);
-          expect(readMcpOAuthStore(identity.storeKey)).toMatchObject({
-            tokens: { access_token: expect.any(String) },
-          });
-          expect(readMcpOAuthStore(identity.storeKey)).not.toHaveProperty("codeVerifier");
-          expect(readMcpOAuthStore(identity.storeKey)).not.toHaveProperty("authorizationAttempt");
-
-          const secondIdentity = requesterIdentity("fixture", config.url, "sender-b");
-          const second = await startMcpOAuthAuthorization(secondIdentity, config, {});
-          if (second.status !== "redirect") {
-            throw new Error("expected second MCP OAuth redirect");
-          }
-          await expect(
-            completeMcpOAuthAuthorization(secondIdentity, config, { code: "wrong-code" }),
-          ).rejects.toThrow();
-          expect(readMcpOAuthStore(secondIdentity.storeKey)).toMatchObject({
-            lastAuthorizationUrl: second.authorizationUrl,
-            redirectUrl: second.redirectUrl,
-            codeVerifier: expect.any(String),
-          });
-          expect(readMcpOAuthStore(secondIdentity.storeKey)).not.toHaveProperty("tokens");
-
-          const third = await startMcpOAuthAuthorization(secondIdentity, config, {});
-          if (third.status !== "redirect") {
-            throw new Error("expected third MCP OAuth redirect");
-          }
-          expect(third.authorizationUrl).not.toBe(second.authorizationUrl);
-          expect(readPending(second.state)).toBeUndefined();
-          expect(readPending(third.state)).toBe(secondIdentity.storeKey);
-          await expect(
-            completeMcpOAuthAuthorization(secondIdentity, config, {
-              code: authorizationCode(second.authorizationUrl),
-            }),
-          ).rejects.toThrow();
-          expect(readMcpOAuthStore(secondIdentity.storeKey).lastAuthorizationUrl).toBe(
-            third.authorizationUrl,
-          );
-          await expect(
-            completeMcpOAuthAuthorization(secondIdentity, config, {
-              code: authorizationCode(third.authorizationUrl),
-            }),
-          ).resolves.toBe("authorized");
-        } finally {
-          await fixture.close();
-        }
-      },
-      {
-        prefix: "openclaw-mcp-oauth-session-",
-        skipSessionCleanup: true,
-        env: { OPENCLAW_CONFIG_PATH: undefined, OPENCLAW_STATE_DIR: undefined },
       },
     );
   });
